version 2.4.0-alpha

+++ classes/controllers/class-controller-order.php
@@ -23,20 +23,28 @@ class Controller_order extends Controller {
 	}
 
 	public function action_add_comment() {
-		$request = $_REQUEST;
-		$note_id = $this->get('payments')->insert_payment_note($request['order_id'], $request['noteText']);
 
-		$this->date['success'] = (is_numeric($note_id)) ? true : false;
-		if ($this->date['success']) {
-			$this->date['data']['html'] = $this->get('payments')->get_payment_note_html($note_id, $request['order_id']);
+		if ( current_user_can('manage_restaurant_menu') ) {
+
+			$request = $_REQUEST;
+			$note_id = $this->get('payments')->insert_payment_note($request['order_id'], $request['noteText']);
+
+			$this->date['success'] = (is_numeric($note_id)) ? true : false;
+			if ($this->date['success']) {
+				$this->date['data']['html'] = $this->get('payments')->get_payment_note_html($note_id, $request['order_id']);
+			}
+			$this->send_json($this->date);
 		}
-		$this->send_json($this->date);
 	}
 
 	public function action_remove_comment() {
-		$request = $_REQUEST;
-		$this->date['success'] = $this->get('payments')->delete_payment_note($request['note_id'], $request['order_id']);
-		$this->send_json($this->date);
+
+		if ( current_user_can('manage_restaurant_menu') ) {
+
+			$request = $_REQUEST;
+			$this->date['success'] = $this->get('payments')->delete_payment_note($request['note_id'], $request['order_id']);
+			$this->send_json($this->date);
+		}
 	}
 
 }
